+++ HISTORY.rst
@@ -1,3 +1,10 @@
+0.4 (2014-11-27)
+----------------
+
+* auto configuration of ALLOWED_HOSTS
+* regex domain matching support
+
+
 0.3 (2014-11-25)
 ----------------